Do not use prepared statements for bulk inserts.
An INSERT INTO that creates N rows is expanded into the following
query:

  INSERT INTO table(...) VALUES ($1,$2,...), ($K,$K+1,...),...
                                ------------------------------
                                           N tuples

This way, we get a unique query for every value of N. This overflows
the cache of prepared statements quickly. Just do not use prepared
statements when doing bulk inserts.

Postgres has a better solution with its COPY protocol. It allows us
to use one prepared statement for N = 1 to insert multiple values.
Unfortunately, I have found no good way to use the COPY protocol
when an insert has an ON CONFLICT clause. Let us stick with a simpler
solution for now.